# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GlowBook - Multi-vendor Hair Salon Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── USERS ────────────────────────────────────────────────────────────────
...
  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: Get current user's bookings
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, in_progress, completed, cancelled]
        - name: role
          in: query
          schema:
            type: string
            enum: [customer, owner]
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"

    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (owner/user)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # ─── REVIEWS ──────────────────────────────────────────────────────────────
  /salons/{salonId}/reviews:
    get:
      operationId: getSalonReviews
      tags: [reviews]
      summary: Get reviews for a salon
      parameters:
        - name: salonId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a salon
      security:
        - bearerAuth: []
      parameters:
        - name: salonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, owner]
          default: user
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ─── USER ─────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, owner, admin]
        createdAt:
          type: string
          format: date-time
...
    Salon:
      type: object
      required: [id, name, ownerId, address, city, isActive, avgRating, totalReviews, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        ownerId:
          type: integer
        address:
          type: string
        city:
          type: string
        state:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        phone:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        avgRating:
          type: number
        totalReviews:
          type: integer
        isActive:
          type: boolean
        isVerified:
          type: boolean
        openTime:
          type: ["string", "null"]
        closeTime:
          type: ["string", "null"]
        distanceKm:
          type: ["number", "null"]
        isFavorited:
          type: ["boolean", "null"]
        createdAt:
          type: string
          format: date-time
...
    Service:
      type: object
      required: [id, salonId, name, price, durationMinutes, isActive, createdAt]
      properties:
        id:
          type: integer
        salonId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        durationMinutes:
          type: integer
        category:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        discountPercent:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
...
    StaffMember:
      type: object
      required: [id, salonId, name, isAvailable, createdAt]
      properties:
        id:
          type: integer
        salonId:
          type: integer
        name:
          type: string
        role:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Booking:
      type: object
      required: [id, userId, salonId, serviceId, bookingDate, startTime, status, totalPrice, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        salonId:
          type: integer
        serviceId:
          type: integer
        staffId:
          type: ["integer", "null"]
        bookingDate:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, in_progress, completed, cancelled]
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
        salon:
          $ref: "#/components/schemas/Salon"
        service:
          $ref: "#/components/schemas/Service"
        staff:
          $ref: "#/components/schemas/StaffMember"
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
    BookingInput:
      type: object
      required: [salonId, serviceId, bookingDate, startTime]
      properties:
        salonId:
          type: integer
        serviceId:
          type: integer
        staffId:
          type: ["integer", "null"]
        bookingDate:
          type: string
          format: date
        startTime:
          type: string
        notes:
          type: ["string", "null"]
    BookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, in_progress, completed, cancelled]

    # ─── REVIEW ───────────────────────────────────────────────────────────
    Review:
      type: object
      required: [id, userId, salonId, rating, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        salonId:
          type: integer
        bookingId:
          type: ["integer", "null"]
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        bookingId:
          type: ["integer", "null"]
    ReviewListResponse:
      type: object
      required: [reviews, total, page, limit, avgRating]
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        avgRating:
          type: number