# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GlowBook - Multi-vendor Hair Salon Booking Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  salonId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  durationMinutes: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  /** @nullable */
  discountPercent?: number | null;
  createdAt: Date;
}